Hook up control points between our player and Youtube
The short version: we're applying the interface that the mediaplayer()
object this library has uses against the Youtube iFrame API.

Most of these tests are highly evented which is why there are nested
waitsFor and runs blocks. These look dirty but they're the only way we
can satisfy an event has been handled over the network as part of the
request.

The biggest decision made here is to pass as much of the video controls
to the Youtube API. We shouldn't have to worry about much of what's
going on for the functionality to work.