Use a YoutubeDecorator to wrap the YoutubePlayer
Given the way that the 'mediaplayer' object is built, we need a single
point of access which holds both the logic to interact with the video
provider and the display logic.

To make this easier a decorator is being used to wrap the YoutubePlayer.
This will largely hook up the event handlers for the accessible elements
on the page. As well as this it passes through the equivalent set of
methods to the Youtubeplayer using an anonymous function wrap.

This isn't ideal, but the intention is for this decorator (read: object
wrapper) to be a simpler addition to the codebase without requiring
large refactoring of the origin jquery.player.js lib.